KGSL: Simple Governor: so I kinda butchered in a good way of course
The original code from faux, but the simplicity is still here.
Added an history so that we can track the average do that the frequency scaling is much more
proportional to the actual GPU load rather than having is
ramp up every now and then due to spurious load spike. Also remove two useless ifs.

Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>